requirements_tests.txt
click==8.1.7
hydra-core==1.3.2
matplotlib==3.5.1
omegaconf==2.3.0
pytest==7.4.4
pytorch_lightning==2.1.3
scikit_learn==1.3.2
seaborn==0.13.1
torch==2.1.2
